Only rebuild index on load
