+++ src/ServicePulse.Host/app/modules/modules.webpackconfig.js
@@ -2,7 +2,6 @@
 const CleanWebpackPlugin = require('clean-webpack-plugin');
 const webpack = require('webpack');
 const ExtractTextPlugin = require('extract-text-webpack-plugin');
-const fs = require('fs');
 
 module.exports = {
     entry: {
@@ -25,23 +24,6 @@ module.exports = {
             moment: 'moment'
         }),
         new ExtractTextPlugin("vendor.css"),
-        // function to replace src="modules/dist/*.js" with the hash appended
-        function() {
-            this.plugin('done', function (statsData) {
-                const stats = statsData.toJson();
-
-                if (!stats.errors.length) {
-                    const html = fs.readFileSync('./app/index.html', 'utf8');
-                    const now = Date.now();
-
-                    let tokenizedMarkup = html.split('.js?v=').map(token => token.split('"></script>'));
-                    let flattenedTokens = tokenizedMarkup.reduce((acc, val) => acc.concat(val), []);
-                    let htmlOutput = flattenedTokens.filter((token, index) => index % 2 == 0).join('.js?v=' + now + '"></script>');
-
-                    fs.writeFileSync('./app/index.html', htmlOutput);
-                }
-            });
-        }
     ],
     module: {
         rules: [{
